requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
